added correct accuracy display

See merge request ipd-boehm/pse/ipd-pse-2022-ws/subspace-outlier-profiling/team-2/pse-implementation!72

+++ backend/src/execution/experiment_scheduler/__init__.py
@@ -88,8 +88,9 @@ class ExperimentScheduler(ABC):
         """
         if ground_truth is None:
             return
-        correct_predictions = np.sum(ground_truth == experiment_result.result_space.outlier_array)
+        correct_predictions = np.sum(ground_truth.astype(int) == experiment_result.result_space.outlier_array.astype(int))
         experiment_result.accuracy = correct_predictions / ground_truth.size
+
 
     @staticmethod
     def get_subspace(dataset: pd.DataFrame, columns: list[int]):
